Start fixing unit tests
