+++ src/components/viewmodels/calendar.view-state.ts
@@ -1,0 +1,5 @@
+import {CalendarUiModel} from 'src/components/models/calendar.ui-model';
+
+export interface CalendarViewState {
+    uiModel?: CalendarUiModel;
+}
