+++ src/domain/models/day.spec.ts
@@ -5,7 +5,6 @@ describe('dayEquals', () => {
         const dayA: Day | undefined = undefined;
         const dayB: Day = {
             date: new Date(2023, 0, 1),
-            properties: { hasPeriodicNote: false },
             dayOfWeek: DayOfWeek.Monday,
             name: 'Monday',
         };
@@ -17,13 +16,11 @@ describe('dayEquals', () => {
     it('should return false if the dates are different', () => {
         const dayA: Day = {
             date: new Date(2023, 0, 1),
-            properties: { hasPeriodicNote: false },
             dayOfWeek: DayOfWeek.Monday,
             name: 'Monday',
         };
         const dayB: Day = {
             date: new Date(2023, 0, 2),
-            properties: { hasPeriodicNote: false },
             dayOfWeek: DayOfWeek.Tuesday,
             name: 'Tuesday',
         };
@@ -34,13 +31,11 @@ describe('dayEquals', () => {
     it('should return true if the dates are the same', () => {
         const dayA: Day = {
             date: new Date(2023, 0, 1),
-            properties: { hasPeriodicNote: false },
             dayOfWeek: DayOfWeek.Monday,
             name: 'Monday',
         };
         const dayB: Day = {
             date: new Date(2023, 0, 1),
-            properties: { hasPeriodicNote: false },
             dayOfWeek: DayOfWeek.Monday,
             name: 'Monday',
         };
