+++ src/implementation/events/select-day.event.spec.ts
@@ -9,9 +9,8 @@ describe('SelectDayNoteEvent', () => {
         event = new SelectDayEvent();
         day = {
             dayOfWeek: DayOfWeek.Tuesday,
-            date: 12,
+            date: new Date('2024-11-12'),
             name: '12',
-            completeDate: new Date('2024-11-12')
         };
     });
