+++ src/implementation/repositories/day.note-repository.spec.ts
@@ -16,7 +16,7 @@ describe('DayNoteRepository', () => {
     });
 
     it('should call getNotesCreatedOn on the note adapter with the correct date', async () => {
-        const day: Day = { dayOfWeek: 1, date: 2, completeDate: new Date(2023, 9, 2), name: '2' };
+        const day: Day = { dayOfWeek: 1, date: new Date(2023, 9, 2), name: '2' };
         const notes: Note[] = [
             { name: 'Note 1', createdOn: new Date(2023, 9, 2, 10), path: 'path/to/note1' },
             { name: 'Note 2', createdOn: new Date(2023, 9, 2, 11), path: 'path/to/note2' },
@@ -26,18 +26,18 @@ describe('DayNoteRepository', () => {
 
         const result = await repository.getNotesCreatedOn(day);
 
-        expect(mockNoteAdapter.getNotesCreatedOn).toHaveBeenCalledWith(day.completeDate);
+        expect(mockNoteAdapter.getNotesCreatedOn).toHaveBeenCalledWith(day.date);
         expect(result).toEqual(notes);
     });
 
     it('should return an empty array when no notes are found', async () => {
-        const day: Day = { dayOfWeek: 1, date: 2, completeDate: new Date(2023, 9, 2), name: '2' };
+        const day: Day = { dayOfWeek: 1, date: new Date(2023, 9, 2), name: '2' };
 
         (mockNoteAdapter.getNotesCreatedOn as jest.Mock).mockResolvedValueOnce([]);
 
         const result = await repository.getNotesCreatedOn(day);
 
-        expect(mockNoteAdapter.getNotesCreatedOn).toHaveBeenCalledWith(day.completeDate);
+        expect(mockNoteAdapter.getNotesCreatedOn).toHaveBeenCalledWith(day.date);
         expect(result).toEqual([]);
     });
 });
